requirements.txt
letta==0.6.50
python_dotenv
openai
